Removed matrix class and fixed bugs
Matrix classes was removed as it was unnecessarily complex. Classed was
replaced by a function.

Moved the traceroute test result retrieval from the Matrix class into
the TracerouteAnalysis class located in classes/traceroute/analysis.py.

Fixed route_cleaner from failing when it received different RTT and AS
values.

Added the ability to retrieve traceroute tests from multiple perfSONAR
Measurement Archives
